Web browser shortcut: Launch Chromium if Firefox not available
The rationale for this change is that Eclipse is highly crash-prone when
Firefox 4 is installed, so before performing certain actions in Eclipse,
Firefox must be uninstalled. At least this way, I can still use the
"launch browser" key combo during my frequently irritating sessions with
Eclipse, using Chromium as a fallback.